kubernetes-chat-app/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
